requirements.txt
Flask
pdfplumber
requests
python-dotenv
google-generativeai
Flask
pandas
numpy
torch
torchvision
torchaudio
scipy
SQLAlchemy
psycopg2-binary
python-dotenv
